Add ability to inject new/rewritten packets
Modify the library to include the ability to inject a new or modified
packet in the place of the original packet when setting a verdict.

Replacing a packet is performed with a new method:

SetVerdictWithPacket(v Verdict, packet []bytes)

This patch includes all modifications to both the Go and C sides to
make this work.

This work is based on the patches found in the (out of date) GitHub
repo at:

https://github.com/david415/go-netfilter-queue

Signed-off-by: William Katsak <bill@dcntral.com>