Removed golang verdict containers and channels
Instead the information needed to call nfq_set_verdict() is now stored
in teh NFPacket struct, meaning that we can now call nfq_set_verdict()
from golang at any point, instead of calling it from the C side after
the golang callback returned.

This also means that the golang callback can return before the decision for a packet is known, which is usefull for handling the packets in multiple threads.